test: organize Go test layout
