+++ internal/config/eslint_plugin_test.go
@@ -7,7 +7,33 @@ import (
 	"github.com/web-infra-dev/rslint/internal/rule"
 )
 
+func preserveGlobalRuleRegistryEntries(t *testing.T, names ...string) {
+	t.Helper()
+
+	type savedRule struct {
+		value  rule.Rule
+		exists bool
+	}
+	saved := make(map[string]savedRule, len(names))
+	for _, name := range names {
+		value, exists := GlobalRuleRegistry.GetRule(name)
+		saved[name] = savedRule{value: value, exists: exists}
+	}
+
+	t.Cleanup(func() {
+		for name, entry := range saved {
+			if entry.exists {
+				GlobalRuleRegistry.rules[name] = entry.value
+			} else {
+				delete(GlobalRuleRegistry.rules, name)
+			}
+		}
+	})
+}
+
 func TestRegisterEslintPluginRules_RegistersPlaceholders(t *testing.T) {
+	preserveGlobalRuleRegistryEntries(t, "testplugA/no-foo", "testplugA/no-bar")
+
 	RegisterEslintPluginRules([]EslintPluginEntry{
 		{Prefix: "testplugA", RuleNames: []string{"no-foo", "no-bar"}},
 	})
@@ -28,6 +54,8 @@ func TestRegisterEslintPluginRules_RegistersPlaceholders(t *testing.T) {
 }
 
 func TestRegisterEslintPluginRules_NativeWins(t *testing.T) {
+	preserveGlobalRuleRegistryEntries(t, "testplugB/native-rule")
+
 	// Pre-register a native rule (IsEslintPluginRule=false), then try to
 	// mount a plugin rule of the same fully-qualified name.
 	GlobalRuleRegistry.Register("testplugB/native-rule", rule.Rule{
@@ -115,6 +143,8 @@ func TestLanguageOptions_RawCaptureAndMerge(t *testing.T) {
 }
 
 func TestGetEnabledRules_PluginGateAndResolution(t *testing.T) {
+	preserveGlobalRuleRegistryEntries(t, "testplugC/no-null")
+
 	RegisterEslintPluginRules([]EslintPluginEntry{
 		{Prefix: "testplugC", RuleNames: []string{"no-null"}},
 	})
@@ -161,6 +191,8 @@ func TestGetEnabledRules_PluginGateAndResolution(t *testing.T) {
 // rules, each carrying the correct IsEslintPluginRule routing flag (native runs
 // in Go, community routes to the worker).
 func TestGetEnabledRules_SplitEntryNativeAndCommunity(t *testing.T) {
+	preserveGlobalRuleRegistryEntries(t, "testplugSplit/no-foo")
+
 	RegisterAllRules()
 	RegisterEslintPluginRules([]EslintPluginEntry{
 		{Prefix: "testplugSplit", RuleNames: []string{"no-foo"}},
@@ -202,6 +234,8 @@ func TestGetEnabledRules_SplitEntryNativeAndCommunity(t *testing.T) {
 // Config resolution retains native and community rules independently of source
 // capabilities. The linter plan owns type-aware eligibility.
 func TestGetEnabledRulesKeepsNativeAndCommunityRules(t *testing.T) {
+	preserveGlobalRuleRegistryEntries(t, "unicornGap/no-null")
+
 	RegisterAllRules()
 	RegisterEslintPluginRules([]EslintPluginEntry{
 		{Prefix: "unicornGap", RuleNames: []string{"no-null"}},
