+++ internal/linter/lint_plan_test.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/microsoft/typescript-go/shim/ast"
+	"github.com/microsoft/typescript-go/shim/compiler"
 	"github.com/web-infra-dev/rslint/internal/plugins/import/rules/no_cycle"
 	lintprogram "github.com/web-infra-dev/rslint/internal/program"
 	"github.com/web-infra-dev/rslint/internal/rule"
@@ -822,4 +823,21 @@ func TestPreparedLintPlanFreezesProgramTypeCapability(t *testing.T) {
 
 func TestLintSingleFileWithoutRuleHandlerIsNoOp(t *testing.T) {
 	LintSingleFile(LintSingleFileOptions{})
+}
+
+func mustSourceOnlyTestProgram(t testing.TB, typeScript *compiler.Program, files []*ast.SourceFile) *lintprogram.Program {
+	t.Helper()
+	sourceProgram, err := lintprogram.NewFromBoundSources(typeScript, files)
+	if err != nil {
+		t.Fatalf("NewFromBoundSources: %v", err)
+	}
+	return sourceProgram
+}
+
+func wrapTestPrograms(programs ...*compiler.Program) []*lintprogram.Program {
+	return lintprogram.NewFromCompilers(programs)
+}
+
+func testPrograms(programs ...*lintprogram.Program) []*lintprogram.Program {
+	return programs
 }
